+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 NOT All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
--Nothing changed just test of release-it
+
+
 
 ## [1.7.2-0] - 2025-08-04
 ## [1.7.1-1] - 2025-08-04