+++ CHANGELOG.md
@@ -4,7 +4,17 @@ NOT All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### ADD
 
+-automatic camera boundres
+
+### REMOVED
+
+-suport for manual boundres
+
+### FIX
+
+-x camera is now working in every scale properly
 
 ## [1.7.2-0] - 2025-08-04
 ## [1.7.1-1] - 2025-08-04