+++ CHANGELOG.md
@@ -15,6 +15,7 @@ NOT All notable changes to this project will be documented in this file.
 ### FIX
 
 -x camera is now working in every scale properly
+-y camera is now working in every scale properly
 
 ## [1.7.2-0] - 2025-08-04
 ## [1.7.1-1] - 2025-08-04